requirements.txt
# Core dependencies
python-telegram-bot>=20.0
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0

# Image processing
opencv-python>=4.8.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.28.0
